+++ src/menu/mod.rs
@@ -222,6 +222,18 @@ impl MenuManager {
         action_map.insert(toggle_fullscreen.id().clone(), MenuAction::ToggleFullscreen);
         view_menu.append(&toggle_fullscreen)?;
 
+        let maximize_vertically = MenuItem::with_id(
+            "maximize_vertically",
+            "Maximize Vertically",
+            true,
+            Some(Accelerator::new(Some(Modifiers::SHIFT), Code::F11)),
+        );
+        action_map.insert(
+            maximize_vertically.id().clone(),
+            MenuAction::MaximizeVertically,
+        );
+        view_menu.append(&maximize_vertically)?;
+
         view_menu.append(&PredefinedMenuItem::separator())?;
 
         let increase_font = MenuItem::with_id(